VFS-932 Fix compilation errors.

+++ include/events/eventAggregator.h
@@ -55,8 +55,8 @@ public:
         virtual std::shared_ptr<Event> processEvent(std::shared_ptr<Event> event, NumericProperty threshold,  const std::string & fieldName, const std::string & sumFieldName);
 
     private:
-        NumericProperty m_counter = 0;			        ///< Aggregated value
-        boost::icl::interval_set<long long> m_blocks;   ///< Aggregated blocks
+        NumericProperty m_counter = 0;			  ///< Aggregated value
+        boost::icl::interval_set<off_t> m_blocks; ///< Aggregated blocks
         std::mutex m_aggregatorStateMutex;
 
         void resetState();					 ///< Resets state - should be called after every event forward.
